requirements.txt
loguru
aiogram
python-dotenv
tinydb
aiohttp_socks
parse